[HIB-188] Hyp3 Daemon gets config passed in
