+++ modules/hyp3_process/src/hyp3_process/hyp3_process.py
@@ -1,38 +1,20 @@
-from typing import Dict, NamedTuple, Callable, Union
-import functools
+from typing import Union
 
-import asf_granule_util as gu
-from hyp3_events import StartEvent
-
-from . import working_directory
-from .outputs import OutputPatterns
-from . import package
-from . import products
-
-
-class EarthdataCredentials(NamedTuple):
-    username: str
-    password: str
-
-
-HandlerFunction = Callable[[
-    StartEvent,
+from .hyp3_handler import (
+    hyp3_handler,
     EarthdataCredentials,
-    str
-], Dict[str, str]
-]
+    HandlerFunction,
+    ProcessingFunction
+)
 
 
 class Process:
     def __init__(
-        self,
-        earthdata_creds: EarthdataCredentials,
-        products_bucket: str
+        self
     ) -> None:
-        self.earthdata_creds = earthdata_creds
-        self.products_bucket = products_bucket
-
-        self.process_handler: Union[HandlerFunction, None] = None
+        self.process_handler: Union[ProcessingFunction, None] = None
+        self.earthdata_creds = {}
+        self.products_bucket = None
 
     def handler(self, process_func: HandlerFunction):
         if self.process_handler is not None:
@@ -48,40 +30,6 @@ class Process:
             self.earthdata_creds,
             self.products_bucket
         )
-
-
-def hyp3_handler(handler_function) -> HandlerFunction:
-    def hyp3_wrapper(
-            job: StartEvent,
-            earthdata_creds: EarthdataCredentials,
-            products_bucket: str
-    ) -> Dict[str, str]:
-        granule = gu.SentinelGranule(job.granule)
-
-        with working_directory.create(granule) as working_dir:
-            gu.download(granule, earthdata_creds, directory=str(working_dir))
-
-            handler_function(granule, working_dir, job.script_path)
-
-            patterns = OutputPatterns(**job.output_patterns)
-
-            process_outputs = package.outputs(
-                archive_name=f'{granule}-rtc-snap',
-                working_dir=working_dir,
-                output_patterns=patterns
-            )
-
-            product_zip_url, browse_url = products.upload(
-                outputs=process_outputs,
-                bucket_name=products_bucket
-            )
-
-        return {
-            'product_url': product_zip_url,
-            'browse_url': browse_url
-        }
-
-    return hyp3_wrapper
 
 
 class HandlerRedefinitionError(Exception):
